`CatalogLegend.vue` → `CatalogTree.vue`
